Updates & Bugfixes
readme.md
-updated links and instructions

Unity Package
- updated preMaterials.cs to pull additional scaling information from materials

Blender Addon
- updated __init__.py to automatically toggle lightmaps off before gltf export
- updated fixMappingNodesRemoveExtras.py to extract scaling data from MaterialsCheckReport.txt
- updated getLightmapInformation to use a 16 digit string in lightmap material name, avoiding duplicate lightmap names between glb files
- minor fixes and tweaks to files
