+++ blender/UnityToVircadia/__init__.py
@@ -32,6 +32,29 @@ class ToggleLightmapsOperator(bpy.types.Operator):
         context.scene.lightmaps_applied = not context.scene.lightmaps_applied
         return {'FINISHED'}
 
+class AdjustShadersOperator(bpy.types.Operator):
+    """Adjust Shaders"""
+    bl_idname = "object.adjust_shaders"
+    bl_label = "Adjust Shaders"
+    bl_description = "Adjust shaders based on the MaterialsCheckReport.txt file"
+
+    def execute(self, context):
+        bpy.ops.object.adjust_shaders('INVOKE_DEFAULT')
+        return {'FINISHED'}
+
+class ExportGLTFOperator(bpy.types.Operator):
+    """Export glTF 2.0"""
+    bl_idname = "export_scene.unity_to_vircadia_gltf"
+    bl_label = "Export glTF 2.0"
+    bl_description = "Export the scene as glTF 2.0 after clearing lightmaps if necessary"
+
+    def execute(self, context):
+        if context.scene.lightmaps_applied:
+            bpy.ops.object.toggle_lightmaps()
+        
+        bpy.ops.export_scene.gltf('INVOKE_DEFAULT')
+        return {'FINISHED'}
+
 class UnityToVircadiaPanel(bpy.types.Panel):
     bl_label = "UnityToVircadia"
     bl_idname = "OBJECT_PT_UnityToVircadia"
@@ -45,7 +68,7 @@ class UnityToVircadiaPanel(bpy.types.Panel):
         layout.operator("import_scene.gltf", text="Import glTF 2.0")
         # layout.operator("object.correct_scale") # TODO later
         layout.operator("import.lightmap_info")
-        layout.operator("object.adjust_shaders")
+        layout.operator("object.adjust_shaders", text="Adjust Shaders")
 
         # Toggle button for applying/clearing lightmaps
         if context.scene.lightmaps_applied:
@@ -53,10 +76,12 @@ class UnityToVircadiaPanel(bpy.types.Panel):
         else:
             layout.operator("object.toggle_lightmaps", text="Preview Lightmaps")
 
-        layout.operator("export_scene.gltf", text="Export glTF 2.0")
+        layout.operator("export_scene.unity_to_vircadia_gltf", text="Export glTF 2.0")
 
 def register():
     bpy.utils.register_class(ToggleLightmapsOperator)
+    bpy.utils.register_class(AdjustShadersOperator)
+    bpy.utils.register_class(ExportGLTFOperator)
     bpy.utils.register_class(UnityToVircadiaPanel)
     bpy.utils.register_class(makeObjectsSingleUserFixScale.CorrectScale)
     bpy.utils.register_class(getLightmapInformation.GetLightmapInfo)
@@ -65,6 +90,8 @@ def register():
 
 def unregister():
     bpy.utils.unregister_class(ToggleLightmapsOperator)
+    bpy.utils.unregister_class(AdjustShadersOperator)
+    bpy.utils.unregister_class(ExportGLTFOperator)
     bpy.utils.unregister_class(UnityToVircadiaPanel)
     bpy.utils.unregister_class(makeObjectsSingleUserFixScale.CorrectScale)
     bpy.utils.unregister_class(getLightmapInformation.GetLightmapInfo)
